% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_general_information.R
\name{historical_general_information}
\alias{historical_general_information}
\title{Get historical information for investing funds}
\usage{
historical_general_information(
  fund_type = c("YAT", "EMK", "BYF"),
  start_date,
  end_date,
  fund_code = NULL
)
}
\arguments{
\item{fund_type}{Type of the fund. Must be one of the following:
\itemize{
\item \code{"YAT"} : Securities Mutual Funds
\item \code{"EMK"} : Pension Funds
\item \code{"BYF"} : Exchange Traded Funds
}}

\item{start_date}{Start date for the historical data. Must be a Date object.}

\item{end_date}{End date for the historical data. Must be a Date object.}

\item{fund_code}{(optional) Abbreviation for the desired fund. If \code{NULL}, it returns all of the funds for the given period.}
}
\value{
A data frame.
}
\description{
Get historical information for investing funds
}
\examples{
historical_general_information(
    fund_type = "YAT",
    start_date = as.Date('2021-10-01'),
    end_date = as.Date('2021-10-31')
)
}
